DOMA-2645 remove ajv to another file as it fails to build
