+++ apps/condo/domains/billing/schema/fields/BillingReceipt/Recipient.js
@@ -1,6 +1,7 @@
 const { render } = require('@condo/domains/billing/schema/fields/utils/json.utils')
 const { BILLING_RECEIPT_RECIPIENT_FIELD_NAME, BILLING_RECEIPT_RECIPIENT_INPUT_NAME } = require('@condo/domains/billing/constants/constants')
 const { RECIPIENT_FIELDS_DEFINITION, RECIPIENT_FIELD: COMMON_RECIPIENT_FIELD, RECIPIENT_QUERY_LIST } = require('@condo/domains/common/schema/fields.js')
+const { RECIPIENT_VALIDATE_HOOK } = require('@condo/domains/common/schema/json.validators')
 
 const RECIPIENT_GRAPHQL_TYPES = `
     type ${BILLING_RECEIPT_RECIPIENT_FIELD_NAME} {
@@ -18,6 +19,9 @@ const RECIPIENT_FIELD = {
     extendGraphQLTypes: [RECIPIENT_GRAPHQL_TYPES],
     graphQLReturnType: BILLING_RECEIPT_RECIPIENT_FIELD_NAME,
     graphQLInputType: BILLING_RECEIPT_RECIPIENT_INPUT_NAME,
+    hooks: {
+        validateInput: RECIPIENT_VALIDATE_HOOK,
+    },
 }
 
 module.exports = {
